feat: add support for node
