+++ packages/mocks/plugins.js
@@ -1,0 +1,2 @@
+export { mockRollupPlugin } from './rollup-plugin.js';
+export { mockPlugin } from './wds-plugin.js';
